kafka-producer/requirements.txt
python-dotenv==1.0.1
confluent-kafka==2.3.0